+++ packages/spec/__testFixtures__/eslint/index.ts
@@ -1,0 +1,4 @@
+interface Foo{
+  a: string,
+  b: number,
+}