+++ packages/spec/src/eslint/common.ts
@@ -1,0 +1,9 @@
+import type { Linter } from 'eslint';
+
+const commonESLintConfig: Linter.Config = {
+  extends: [
+    '@applint/eslint-config',
+  ],
+};
+
+export default commonESLintConfig;