+++ packages/spec/src/eslint/rax-ts.js
@@ -1,7 +1,0 @@
-const getRaxESlintConfig = require('../getRaxESlintConfig');
-
-module.exports = getRaxESlintConfig({
-  extends: [
-    require.resolve('@applint/eslint-config/typescript/rax'),
-  ],
-});