+++ packages/spec/src/eslint/react.ts
@@ -1,0 +1,9 @@
+import type { Linter } from 'eslint';
+
+const reactESLintConfig: Linter.Config = {
+  extends: [
+    '@applint/eslint-config/react',
+  ],
+};
+
+export default reactESLintConfig;