+++ packages/spec/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @applint/spec
 
+## 1.0.0
+
+- refactor: transform language from JS to TS
+- feat: update `@applint/eslint-config` and `@applint/stylelint-config`
+
 ## 0.1.0
 
 - feat: init eslint/stylelint/prettier/commitlint rules.